Use abstract class for master clients

+++ common/src/thrift/tachyon.thrift
@@ -191,11 +191,11 @@ service FileSystemMasterService {
   // TODO: is this necessary?
   string getUfsAddress()
 
-  i32 createFile(1: i64 fileId, 2: i64 blockSizeBytes, 3: bool recursive)
+  i64 createFile(1: i64 fileId, 2: i64 blockSizeBytes, 3: bool recursive)
     throws (1: FileAlreadyExistException faee, 2: BlockInfoException bie,
       3: SuspectedFileSizeException sfse, 4: TachyonException te)
 
-  i32 loadFileFromUfs(1: i64 fileId, 2: string ufsPath, 3: bool recursive)
+  i64 loadFileFromUfs(1: i64 fileId, 2: string ufsPath, 3: bool recursive)
     throws (1: FileAlreadyExistException faee, 2: BlockInfoException bie,
       3: SuspectedFileSizeException sfse, 4: TachyonException te)
